Rename file for translateHttpLoader
